Shrink the sizes of some constants
This commit changes several constants to use smaller types. This means
we don't have to downcast them to use them, and we can safely convert
them to larger types when needed.

One limitation with this is that Kani doesn't support casts in length
fields, so I changed those cases to use a separate constant containing
the cast.

+++ src/seccompiler/src/common.rs
@@ -7,7 +7,7 @@
 use serde::{Deserialize, Serialize};
 
 /// The maximum seccomp-BPF program length allowed by the linux kernel.
-pub const BPF_MAX_LEN: usize = 4096;
+pub const BPF_MAX_LEN: u16 = 4096;
 
 /// BPF instruction structure definition.
 /// See /usr/include/linux/filter.h .
