+++ src/vmm/src/mmds/persist.rs
@@ -18,7 +18,7 @@ use crate::mmds::data_store::Mmds;
 /// State of a MmdsNetworkStack.
 #[derive(Debug, Clone, Versionize)]
 pub struct MmdsNetworkStackState {
-    mac_addr: [u8; MAC_ADDR_LEN],
+    mac_addr: [u8; MAC_ADDR_LEN as usize],
     ipv4_addr: u32,
     tcp_port: u16,
     max_connections: usize,
@@ -31,7 +31,7 @@ impl Persist<'_> for MmdsNetworkStack {
     type Error = ();
 
     fn save(&self) -> Self::State {
-        let mut mac_addr = [0; MAC_ADDR_LEN];
+        let mut mac_addr = [0; MAC_ADDR_LEN as usize];
         mac_addr.copy_from_slice(self.mac_addr.get_bytes());
 
         MmdsNetworkStackState {
